+++ src/medstime/site_index.go
@@ -11,7 +11,7 @@ func indexGet(ctx *web.Context) string {
 
     m := map[string]interface{}{}
     
-    acc_id := ses.GetInt64("account_id")
+    acc_id := ses.GetInt("account_id")
     if acc_id > 0 {
         m["Accid"] = acc_id
     }
