+++ src/shared/account.go
@@ -4,5 +4,5 @@ type Account struct {
 	Password string
 	Email    string
 	Plan     int
-	Id       int64
+	Id       int
 }
