+++ src/shared/schedule.go
@@ -4,16 +4,16 @@ import "fmt"
 
 //a schedule plan - a account has one
 type Schedule struct {
-	Id        int64
-	AccountId int64 //the parent account
+	Id        int
+	AccountId int //the parent account
 }
 
 //an item in the schedule plan
 type ScheduleItem struct {
-	Id                 int64
-	ScheduleId         int64  //the parent schedule
+	Id                 int
+	ScheduleId         int  //the parent schedule
 	Message            string //the message to be sent
-	OffsetFromMidnight int64  //offset from midnight in seconds (to determine fire time)
+	OffsetFromMidnight int  //offset from midnight in seconds (to determine fire time)
 }
 
 //helper methods for mustache formatting - are called by mustche
@@ -28,7 +28,7 @@ func (itm ScheduleItem) Minute() string {
 }
 
 //get number of seconds from midnight for a given hour:minute pair
-func SecondsFromMidnight(hour, minute int64) int64 {
+func SecondsFromMidnight(hour, minute int) int {
 	seconds := minute*60 + hour*(60*60)
 	return seconds
 }
